+++ src/plugins/DistanceMeasurementsPlugin/DistanceMeasurementsControl.js
@@ -152,11 +152,14 @@ class DistanceMeasurementsControl extends Component {
         const input = scene.input;
         const startDot = this._touchStartDot;
 
-        const pickSurfacePrecisionEnabled = scene.pickSurfacePrecisionEnabled;
-
         let mouseHoverEntity = null;
-        const mouseWorldPos = math.vec3();
-        const mouseCanvasPos = math.vec2();
+
+        const surfaceWorldPos = math.vec3();
+        const surfaceCanvasPos = math.vec2();
+
+        const snappedWorldPos = math.vec3();
+        const snappedCanvasPos = math.vec2();
+
 
         let lastMouseCanvasX;
         let lastMouseCanvasY;
@@ -171,13 +174,17 @@ class DistanceMeasurementsControl extends Component {
         const touchEndCanvasPos = math.vec2();
         const touchStartWorldPos = math.vec3();
 
+        let snapped = false;
+
         this._onMouseHoverSurface = cameraControl.on("hoverOverEntitySurface", event => {
+
+            snapped = false;
 
             // This gets fired for both mouse and touch input, but we don't care when handling touch
             mouseHoverEntity = true;
 
-            mouseWorldPos.set(event.worldPos);
-            mouseCanvasPos.set(event.canvasPos);
+            surfaceWorldPos.set(event.worldPos);
+            surfaceCanvasPos.set(event.canvasPos);
 
             if (touchState === FIRST_TOUCH_EXPECTED) {
                 this.markerDiv.style.marginLeft = `${event.canvasPos[0] - 5}px`;
@@ -201,18 +208,21 @@ class DistanceMeasurementsControl extends Component {
                 this._currentDistanceMeasurementByMouse.yAxisVisible = this._currentDistanceMeasurementByMouseInittouchState.yAxisVisible && this.plugin.defaultYAxisVisible;
                 this._currentDistanceMeasurementByMouse.zAxisVisible = this._currentDistanceMeasurementByMouseInittouchState.zAxisVisible && this.plugin.defaultZAxisVisible;
                 this._currentDistanceMeasurementByMouse.targetVisible = this._currentDistanceMeasurementByMouseInittouchState.targetVisible;
-                //   this._currentDistanceMeasurementByMouse.target.entity = mouseHoverEntity;
-                this._currentDistanceMeasurementByMouse.target.worldPos = mouseWorldPos;
+                this._currentDistanceMeasurementByMouse.target.worldPos = surfaceWorldPos.slice();
             }
         });
 
         this._onMouseHoverSurface = cameraControl.on("pickedVertex", event => {
 
+            console.log("pickedVertex");
+
+            snapped = true;
+
             // This gets fired for both mouse and touch input, but we don't care when handling touch
             mouseHoverEntity = true;
 
-            mouseWorldPos.set(event.snappedWorldPos);
-            mouseCanvasPos.set(event.snappedCanvasPos);
+            snappedWorldPos.set(event.snappedWorldPos);
+            snappedCanvasPos.set(event.snappedCanvasPos);
 
             if (touchState === FIRST_TOUCH_EXPECTED) {
                 this.markerDiv.style.marginLeft = `${event.snappedCanvasPos[0] - 5}px`;
@@ -223,8 +233,8 @@ class DistanceMeasurementsControl extends Component {
             }
             // } else {
             //     if (event.worldPos !== null && event.canvasPos !== null) {
-            //         mouseWorldPos.set(event.worldPos);
-            //         mouseCanvasPos.set(event.canvasPos);
+            //         surfaceWorldPos.set(event.worldPos);
+            //         surfaceCanvasPos.set(event.canvasPos);
             //
             //         if (touchState === FIRST_TOUCH_EXPECTED) {
             //             this.markerDiv.style.marginLeft = `${event.canvasPos[0]-5}px`;
@@ -250,25 +260,28 @@ class DistanceMeasurementsControl extends Component {
                 this._currentDistanceMeasurementByMouse.yAxisVisible = this._currentDistanceMeasurementByMouseInittouchState.yAxisVisible && this.plugin.defaultYAxisVisible;
                 this._currentDistanceMeasurementByMouse.zAxisVisible = this._currentDistanceMeasurementByMouseInittouchState.zAxisVisible && this.plugin.defaultZAxisVisible;
                 this._currentDistanceMeasurementByMouse.targetVisible = this._currentDistanceMeasurementByMouseInittouchState.targetVisible;
-                //   this._currentDistanceMeasurementByMouse.target.entity = mouseHoverEntity;
-                this._currentDistanceMeasurementByMouse.target.worldPos = mouseWorldPos;
+                this._currentDistanceMeasurementByMouse.target.worldPos = surfaceWorldPos;
             }
         });
 
         this._onInputMouseDown = input.on("mousedown", (coords) => {
+            console.log("mousedown");
             lastMouseCanvasX = coords[0];
             lastMouseCanvasY = coords[1];
         });
 
         this._onInputMouseUp = input.on("mouseup", (coords) => {
+            console.log("mouseup snapped = " + snapped);
             if (coords[0] > lastMouseCanvasX + mouseCanvasClickTolerance ||
                 coords[0] < lastMouseCanvasX - mouseCanvasClickTolerance ||
                 coords[1] > lastMouseCanvasY + mouseCanvasClickTolerance ||
                 coords[1] < lastMouseCanvasY - mouseCanvasClickTolerance) {
                 return;
             }
+            const worldPos = snapped ? snappedWorldPos.slice() : surfaceWorldPos.slice();
             if (this._currentDistanceMeasurementByMouse) {
                 if (mouseHoverEntity) {
+                    this._currentDistanceMeasurementByMouse.target.worldPos = worldPos;
                     this._currentDistanceMeasurementByMouse.clickable = true;
                     this.fire("measurementEnd", this._currentDistanceMeasurementByMouse);
                     this._currentDistanceMeasurementByMouse = null;
@@ -282,10 +295,10 @@ class DistanceMeasurementsControl extends Component {
                     this._currentDistanceMeasurementByMouse = plugin.createMeasurement({
                         id: math.createUUID(),
                         origin: {
-                            worldPos: mouseWorldPos
+                            worldPos: worldPos.slice()
                         },
                         target: {
-                            worldPos: mouseWorldPos
+                            worldPos: worldPos.slice()
                         },
                         approximate: true
                     });
@@ -304,8 +317,9 @@ class DistanceMeasurementsControl extends Component {
         });
 
         this._onMouseHoverOff = cameraControl.on("hoverNothing", event => {
+            console.log("hoverNothing");
             mouseHoverEntity = null;
-
+            snapped = false;
             this.markerDiv.style.marginLeft = `-100px`;
             this.markerDiv.style.marginTop = `-100px`;
 
@@ -318,72 +332,73 @@ class DistanceMeasurementsControl extends Component {
         });
 
         this._onPickedNothing = cameraControl.on("pickedNothing", event => {
-            if (this._currentDistanceMeasurementByMouse) {
-                this.fire("measurementCancel", this._currentDistanceMeasurementByMouse);
-                this._currentDistanceMeasurementByMouse.destroy();
-                this._currentDistanceMeasurementByMouse = null;
-            }
-            startDot.setVisible(false);
-            touchState = FIRST_TOUCH_EXPECTED;
+            console.log("pickedNothing");
+            // if (this._currentDistanceMeasurementByMouse) {
+            //     this.fire("measurementCancel", this._currentDistanceMeasurementByMouse);
+            //     this._currentDistanceMeasurementByMouse.destroy();
+            //     this._currentDistanceMeasurementByMouse = null;
+            // }
+            // startDot.setVisible(false);
+            // touchState = FIRST_TOUCH_EXPECTED;
         });
 
-        canvas.addEventListener("touchstart", this._onCanvasTouchStart = (event) => {
-            const touches = event.touches;
-            const changedTouches = event.changedTouches;
-            if (touches.length === 1 && changedTouches.length === 1) {
-                getCanvasPosFromEvent(touches[0], touchStartCanvasPos);
-            }
-        }, {passive: true});
-
-        canvas.addEventListener("touchend", this._onCanvasTouchEnd = (event) => {
-            const touches = event.touches;
-            const changedTouches = event.changedTouches;
-            if (touches.length === 0 && changedTouches.length === 1) {
-                getCanvasPosFromEvent(changedTouches[0], touchEndCanvasPos);
-                if (touchEndCanvasPos[0] > touchStartCanvasPos[0] + touchCanvasClickTolerance ||
-                    touchEndCanvasPos[0] < touchStartCanvasPos[0] - touchCanvasClickTolerance ||
-                    touchEndCanvasPos[1] > touchStartCanvasPos[1] + touchCanvasClickTolerance ||
-                    touchEndCanvasPos[1] < touchStartCanvasPos[1] - touchCanvasClickTolerance) {
-                    return; // User is repositioning the camera or model
-                }
-                const pickResult = scene.pick({
-                    canvasPos: touchEndCanvasPos,
-                    pickSurface: true,
-                    pickSurfacePrecision: pickSurfacePrecisionEnabled
-                });
-                if (pickResult && pickResult.worldPos) {
-                    switch (touchState) {
-                        case FIRST_TOUCH_EXPECTED:
-                            startDot.setVisible(true);
-                            this._touchStartMarker.worldPos = pickResult.worldPos;
-                            touchStartWorldPos.set(pickResult.worldPos);
-                            touchState = SECOND_TOUCH_EXPECTED;
-                            break;
-                        case SECOND_TOUCH_EXPECTED:
-                            startDot.setVisible(false);
-                            this._touchStartMarker.worldPos = pickResult.worldPos;
-                            const measurement = plugin.createMeasurement({
-                                id: math.createUUID(),
-                                origin: {
-                                    worldPos: touchStartWorldPos
-                                },
-                                target: {
-                                    worldPos: pickResult.worldPos
-                                },
-                                approximate: (!pickSurfacePrecisionEnabled)
-                            });
-                            measurement.clickable = true;
-                            touchState = FIRST_TOUCH_EXPECTED;
-                            this.fire("measurementEnd", measurement);
-                            break;
-                    }
-                } else {
-                    startDot.setVisible(false);
-                    touchState = FIRST_TOUCH_EXPECTED;
-                }
-            }
-            //  event.stopPropagation();
-        }, {passive: true});
+        // canvas.addEventListener("touchstart", this._onCanvasTouchStart = (event) => {
+        //     const touches = event.touches;
+        //     const changedTouches = event.changedTouches;
+        //     if (touches.length === 1 && changedTouches.length === 1) {
+        //         getCanvasPosFromEvent(touches[0], touchStartCanvasPos);
+        //     }
+        // }, {passive: true});
+        //
+        // canvas.addEventListener("touchend", this._onCanvasTouchEnd = (event) => {
+        //     const touches = event.touches;
+        //     const changedTouches = event.changedTouches;
+        //     if (touches.length === 0 && changedTouches.length === 1) {
+        //         getCanvasPosFromEvent(changedTouches[0], touchEndCanvasPos);
+        //         if (touchEndCanvasPos[0] > touchStartCanvasPos[0] + touchCanvasClickTolerance ||
+        //             touchEndCanvasPos[0] < touchStartCanvasPos[0] - touchCanvasClickTolerance ||
+        //             touchEndCanvasPos[1] > touchStartCanvasPos[1] + touchCanvasClickTolerance ||
+        //             touchEndCanvasPos[1] < touchStartCanvasPos[1] - touchCanvasClickTolerance) {
+        //             return; // User is repositioning the camera or model
+        //         }
+        //         const pickResult = scene.pick({
+        //             canvasPos: touchEndCanvasPos,
+        //             pickSurface: true,
+        //             pickSurfacePrecision: pickSurfacePrecisionEnabled
+        //         });
+        //         if (pickResult && pickResult.worldPos) {
+        //             switch (touchState) {
+        //                 case FIRST_TOUCH_EXPECTED:
+        //                     startDot.setVisible(true);
+        //                     this._touchStartMarker.worldPos = pickResult.worldPos;
+        //                     touchStartWorldPos.set(pickResult.worldPos);
+        //                     touchState = SECOND_TOUCH_EXPECTED;
+        //                     break;
+        //                 case SECOND_TOUCH_EXPECTED:
+        //                     startDot.setVisible(false);
+        //                     this._touchStartMarker.worldPos = pickResult.worldPos;
+        //                     const measurement = plugin.createMeasurement({
+        //                         id: math.createUUID(),
+        //                         origin: {
+        //                             worldPos: touchStartWorldPos
+        //                         },
+        //                         target: {
+        //                             worldPos: pickResult.worldPos
+        //                         },
+        //                         approximate: (!pickSurfacePrecisionEnabled)
+        //                     });
+        //                     measurement.clickable = true;
+        //                     touchState = FIRST_TOUCH_EXPECTED;
+        //                     this.fire("measurementEnd", measurement);
+        //                     break;
+        //             }
+        //         } else {
+        //             startDot.setVisible(false);
+        //             touchState = FIRST_TOUCH_EXPECTED;
+        //         }
+        //     }
+        //     //  event.stopPropagation();
+        // }, {passive: true});
 
         this._active = true;
     }
