+++ src/viewer/scene/frustumCulling/ViewFrustumCullingMesh.js
@@ -1,0 +1,9 @@
+/**
+ * @abstract
+ */
+export class ViewFrustumCullingNode {
+
+    get numtriangles() {
+    }
+}
+
