+++ src/viewer/scene/frustumCulling/ViewFrustumCullingModel.js
@@ -1,0 +1,9 @@
+/**
+ * @abstract
+ */
+export class ViewFrustumCullingModel {
+
+    get worldMatrix() {
+    }
+
+}
