+++ src/viewer/scene/frustumCulling/ViewFrustumCullingNode.js
@@ -1,0 +1,12 @@
+/**
+ * @abstract
+ */
+export class ViewFrustumCullingNode {
+
+    get culledVCF() {
+    }
+
+    set culledVFC(culledVFC) {
+    }
+}
+
