+++ src/viewer/scene/models/DataTextureSceneModel/DataTextureSceneModel.js
@@ -8,12 +8,14 @@ import {ENTITY_FLAGS} from './lib/ENTITY_FLAGS.js';
 import {utils} from "../../utils.js";
 import {RenderFlags} from "../../webgl/RenderFlags.js";
 
-import {LodCullingManager} from "./lib/layers/trianglesDataTexture/LodCullingManager.js";
-import {ViewFrustumCullingManager} from "./lib/layers/trianglesDataTexture/ViewFrustumCullingManager.js";
+// import {LodCullingManager} from "./lib/layers/trianglesDataTexture/LodCullingManager.js";
+//import {ViewFrustumCullingManager} from "./lib/layers/trianglesDataTexture/ViewFrustumCullingManager.js";
 import {buildEdgeIndices} from "../../math/buildEdgeIndices";
 import {rebucketPositions} from "./lib/layers/trianglesDataTexture/rebucketPositions";
 import {uniquifyPositions} from "./lib/layers/trianglesDataTexture/calculateUniquePositions";
 import {quantizePositions} from "./lib/compression";
+import {LodCullingManager} from "../../lodCulling/LodCullingManager";
+import {ViewFrustumCullingManager} from "../../frustumCulling/ViewFrustumCullingManager";
 
 const tempVec3a = math.vec3();
 const tempMat4 = math.mat4();
@@ -26,11 +28,913 @@ const defaultQuaternion = math.identityQuaternion();
 /**
  * @desc A high-performance data-texture-based model representation for efficient rendering and low memory usage.
  *
+ * Data textures offer a highly efficient method for storing model representations on the GPU. They are much more efficient than the standard vertex buffer object (VBO)
+ * technique used in almost every other graphics library.
+ *
+ * By using data textures to store geometry and materials, a xeokit shader can randomly access data from anywhere within the texture. In contrast, with VBOs, shaders
+ * can only access the primitive they are currently rendering. This allows shaders to fetch contextual information about the primitive being rendered, enabling implementation of many useful algorithms on the GPU, such as LoD, visibility culling, and more.
+ *
+ * Data textures also allow us to use a lot of "pointer indirection" within the textures, combining and reusing duplicate geometry elements and materials. This dramatically
+ * minimizes the GPU memory footprint with large models.
+ *
+ * # Examples
+ *
+ * * [DataTextureSceneModel using geometry batching](/examples/#sceneRepresentation_DataTextureSceneModel_batching)
+ * * [DataTextureSceneModel using geometry batching and RTC coordinates](/examples/#sceneRepresentation_DataTextureSceneModel_batching_origin)
+ * * [DataTextureSceneModel using geometry instancing](/examples/#sceneRepresentation_DataTextureSceneModel_instancing)
+ * * [DataTextureSceneModel using geometry instancing and RTC coordinates](/examples/#sceneRepresentation_DataTextureSceneModel_instancing_origin)
+ *
+ * <br>
+ *
+ * It should be noted that while the terms "batching" and "instancing" are used in the DataTextureModel API, their usage is primarily superficial. They are
+ * employed to ensure the (mostly) seamless implementation of the {@link SceneModel} interface, which is also implemented by {@link VBOSceneModel}. However, internally, the
+ * DataTextureModel adopts its own pointer-based reuse strategy to handle both geometry modes, as previously mentioned.
+ *
+ * # Overview
+ *
+ * While xeokit's standard [scene graph](https://github.com/xeokit/xeokit-sdk/wiki/Scene-Graphs) is great for gizmos and medium-sized models, it doesn't scale up to millions of objects in terms of memory and rendering efficiency.
+ *
+ * For huge models, we have the ````DataTextureSceneModel```` representation, which is optimized to pack large amounts of geometry into memory and render it efficiently using WebGL.
+ *
+ * ````DataTextureSceneModel```` is the default model representation loaded by  (at least) {@link GLTFLoaderPlugin}, {@link XKTLoaderPlugin} and  {@link WebIFCLoaderPlugin}.
+ *
+ * In this tutorial you'll learn how to use ````DataTextureSceneModel```` to create high-detail content programmatically. Ordinarily you'd be learning about ````DataTextureSceneModel```` if you were writing your own model loader plugins.
+ *
+ * # Contents
+ *
+ * - [DataTextureSceneModel](#performancemodel)
+ * - [GPU-Resident Geometry](#gpu-resident-geometry)
+ * - [Picking](#picking)
+ * - [Example 1: Geometry Instancing](#example-1--geometry-instancing)
+ * - [Finalizing a DataTextureSceneModel](#finalizing-a-performancemodel)
+ * - [Finding Entities](#finding-entities)
+ * - [Example 2: Geometry Batching](#example-2--geometry-batching)
+ * - [Classifying with Metadata](#classifying-with-metadata)
+ * - [Querying Metadata](#querying-metadata)
+ * - [Metadata Structure](#metadata-structure)
+ * - [RTC Coordinates](#rtc-coordinates-for-double-precision)
+ *   - [Example 3: RTC Coordinates with Geometry Instancing](#example-2--rtc-coordinates-with-geometry-instancing)
+ *   - [Example 4: RTC Coordinates with Geometry Batching](#example-2--rtc-coordinates-with-geometry-batching)
+ *
+ * ## DataTextureSceneModel
+ *
+ * ````DataTextureSceneModel```` uses two rendering techniques internally:
+ *
+ * 1. ***Geometry batching*** for unique geometries, combining those into a single WebGL geometry buffer, to render in one draw call, and
+ * 2. ***geometry instancing*** for geometries that are shared by multiple meshes, rendering all instances of each shared geometry in one draw call.
+ *
+ * <br>
+ * These techniques come with certain limitations:
+ *
+ * * Non-realistic rendering - while scene graphs can use xeokit's full set of material workflows, ````DataTextureSceneModel```` uses simple Lambertian shading without textures.
+ * * Static transforms - transforms within a ````DataTextureSceneModel```` are static and cannot be dynamically translated, rotated and scaled the way {@link Node}s and {@link Mesh}es in scene graphs can.
+ * * Immutable model representation - while scene graph {@link Node}s and
+ * {@link Mesh}es can be dynamically plugged together, ````DataTextureSceneModel```` is immutable,
+ * since it packs its geometries into textures.
+ *
+ * ````DataTextureSceneModel````'s API allows us to exploit batching and instancing, while exposing its elements as
+ * abstract {@link Entity} types.
+ *
+ * {@link Entity} is the abstract base class for
+ * the various xeokit components that represent models, objects, or anonymous visible elements. An Entity has a unique ID and can be
+ * individually shown, hidden, selected, highlighted, ghosted, culled, picked and clipped, and has its own World-space boundary.
+ *
+ * * A ````DataTextureSceneModel```` is an {@link Entity} that represents a model.
+ * * A ````DataTextureSceneModel```` represents each of its objects with an {@link Entity}.
+ * * Each {@link Entity} has one or more meshes that define its shape.
+ * * Each mesh has either its own unique geometry, or shares a geometry with other meshes.
+ *
+ * ## GPU-Resident Geometry
+ *
+ * For a low memory footprint, ````DataTextureSceneModel```` stores its geometries in GPU memory only, compressed (quantized) as integers. Unfortunately, GPU-resident geometry is
+ * not readable by JavaScript.
+ *
+ *
+ * ## Example 1: Geometry Instancing
+ *
+ * In the example below, we'll use a ````DataTextureSceneModel````
+ * to build a simple table model using geometry instancing.
+ *
+ * We'll start by adding a reusable box-shaped geometry to our ````DataTextureSceneModel````.
+ *
+ * Then, for each object in our model we'll add an {@link Entity}
+ * that has a mesh that instances our box geometry, transforming and coloring the instance.
+ *
+ * [![](http://xeokit.io/img/docs/sceneGraph.png)](/examples/#sceneRepresentation_DataTextureSceneModel_instancing)
+ *
+ * [[Run this example](/examples/#sceneRepresentation_DataTextureSceneModel_instancing)]
+ *
+ * ````javascript
+ * import {Viewer, DataTextureSceneModel} from "xeokit-sdk.es.js";
+ *
+ * const viewer = new Viewer({
+ *     canvasId: "myCanvas",
+ *     transparent: true
+ * });
+ *
+ * viewer.scene.camera.eye = [-21.80, 4.01, 6.56];
+ * viewer.scene.camera.look = [0, -5.75, 0];
+ * viewer.scene.camera.up = [0.37, 0.91, -0.11];
+ *
+ * // Build a DataTextureSceneModel representing a table
+ * // with four legs, using geometry instancing
+ *
+ * const dataTextureSceneModel = new DataTextureSceneModel(viewer.scene, {
+ *     id: "table",
+ *     isModel: true, // <--- Registers DataTextureSceneModel in viewer.scene.models
+ *     position: [0, 0, 0],
+ *     scale: [1, 1, 1],
+ *     rotation: [0, 0, 0]
+ * });
+ *
+ * // Create a reusable geometry within the DataTextureSceneModel
+ * // We'll instance this geometry by five meshes
+ *
+ * dataTextureSceneModel.createGeometry({
+ *
+ *     id: "myBoxGeometry",
+ *
+ *     // The primitive type - allowed values are "points", "lines" and "triangles".
+ *     // See the OpenGL/WebGL specification docs
+ *     // for how the coordinate arrays are supposed to be laid out.
+ *     primitive: "triangles",
+ *
+ *     // The vertices - eight for our cube, each
+ *     // one spanning three array elements for X,Y and Z
+ *     positions: [
+ *          1, 1, 1, -1, 1, 1, -1, -1, 1, 1, -1, 1, // v0-v1-v2-v3 front
+ *          1, 1, 1, 1, -1, 1, 1, -1, -1, 1, 1, -1, // v0-v3-v4-v1 right
+ *          1, 1, 1, 1, 1, -1, -1, 1, -1, -1, 1, 1, // v0-v1-v6-v1 top
+ *          -1, 1, 1, -1, 1, -1, -1, -1, -1, -1, -1, 1, // v1-v6-v7-v2 left
+ *          -1, -1, -1, 1, -1, -1, 1, -1, 1, -1, -1, 1, // v7-v4-v3-v2 bottom
+ *          1, -1, -1, -1, -1, -1, -1, 1, -1, 1, 1, -1 // v4-v7-v6-v1 back
+ *     ],
+ *
+ *     // Normal vectors, one for each vertex
+ *     normals: [
+ *         0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, // v0-v1-v2-v3 front
+ *         1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, // v0-v3-v4-v5 right
+ *         0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, // v0-v5-v6-v1 top
+ *         -1, 0, 0, -1, 0, 0, -1, 0, 0, -1, 0, 0, // v1-v6-v7-v2 left
+ *         0, -1, 0, 0, -1, 0, 0, -1, 0, 0, -1, 0, // v7-v4-v3-v2 bottom
+ *         0, 0, -1, 0, 0, -1, 0, 0, -1, 0, 0, -1 // v4-v7-v6-v5 back
+ *     ],
+ *
+ *     // Indices - these organise the positions and and normals
+ *     // into geometric primitives in accordance with the "primitive" parameter,
+ *     // in this case a set of three indices for each triangle.
+ *     //
+ *     // Note that each triangle is specified in counter-clockwise winding order.
+ *     //
+ *     indices: [
+ *         0, 1, 2, 0, 2, 3, // front
+ *         4, 5, 6, 4, 6, 7, // right
+ *         8, 9, 10, 8, 10, 11, // top
+ *         12, 13, 14, 12, 14, 15, // left
+ *         16, 17, 18, 16, 18, 19, // bottom
+ *         20, 21, 22, 20, 22, 23
+ *     ]
+ * });
+ *
+ * // Red table leg
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "redLegMesh",
+ *     geometryId: "myBoxGeometry",
+ *     position: [-4, -6, -4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [1, 0.3, 0.3],
+ *     origin: cfg.origin
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     id: "redLeg",
+ *     meshIds: ["redLegMesh"],
+ *     isObject: true // <---- Registers Entity by ID on viewer.scene.objects
+ * });
+ *
+ * // Green table leg
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "greenLegMesh",
+ *     geometryId: "myBoxGeometry",
+ *     position: [4, -6, -4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [0.3, 1.0, 0.3],
+ *     origin: cfg.origin
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     id: "greenLeg",
+ *     meshIds: ["greenLegMesh"],
+ *     isObject: true // <---- Registers Entity by ID on viewer.scene.objects
+ * });
+ *
+ * // Blue table leg
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "blueLegMesh",
+ *     geometryId: "myBoxGeometry",
+ *     position: [4, -6, 4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [0.3, 0.3, 1.0],
+ *     origin: cfg.origin
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     id: "blueLeg",
+ *     meshIds: ["blueLegMesh"],
+ *     isObject: true // <---- Registers Entity by ID on viewer.scene.objects
+ * });
+ *
+ * // Yellow table leg
+ *
+ * dataTextureSceneModel.createMesh({
+ *      id: "yellowLegMesh",
+ *      geometryId: "myBoxGeometry",
+ *      position: [-4, -6, 4],
+ *      scale: [1, 3, 1],
+ *      rotation: [0, 0, 0],
+ *      color: [1.0, 1.0, 0.0],
+ *     origin: cfg.origin
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     id: "yellowLeg",
+ *     meshIds: ["yellowLegMesh"],
+ *     isObject: true // <---- Registers Entity by ID on viewer.scene.objects
+ * });
+ *
+ * // Purple table top
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "purpleTableTopMesh",
+ *     geometryId: "myBoxGeometry",
+ *     position: [0, -3, 0],
+ *     scale: [6, 0.5, 6],
+ *     rotation: [0, 0, 0],
+ *     color: [1.0, 0.3, 1.0],
+ *     origin: cfg.origin
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     id: "purpleTableTop",
+ *     meshIds: ["purpleTableTopMesh"],
+ *     isObject: true // <---- Registers Entity by ID on viewer.scene.objects
+ * });
+ *  ````
+ *
+ * ## Finalizing a DataTextureSceneModel
+ *
+ * Before we can view and interact with our ````DataTextureSceneModel````, we need to **finalize** it. Internally, this causes the ````DataTextureSceneModel```` to
+ * build the data textures that support our geometries and materials.
+ *
+ * Once finalized, we can't add anything more to our ````DataTextureSceneModel````.
+ *
+ * ```` javascript
+ * dataTextureSceneModel.finalize();
+ * ````
+ *
+ * ## Finding Entities
+ *
+ * As mentioned earlier, {@link Entity} is
+ * the abstract base class for components that represent models, objects, or just
+ * anonymous visible elements.
+ *
+ * Since we created configured our ````DataTextureSceneModel```` with ````isModel: true````,
+ * we're able to find it as an Entity by ID in ````viewer.scene.models````. Likewise, since
+ * we configured each of its Entities with ````isObject: true````, we're able to
+ * find them in  ````viewer.scene.objects````.
+ *
+ *
+ * ````javascript
+ * // Get the whole table model Entity
+ * const table = viewer.scene.models["table"];
+ *
+ *  // Get some leg object Entities
+ * const redLeg = viewer.scene.objects["redLeg"];
+ * const greenLeg = viewer.scene.objects["greenLeg"];
+ * const blueLeg = viewer.scene.objects["blueLeg"];
+ * ````
+ *
+ * ## Example 2: Geometry Batching
+ *
+ * Let's once more use a ````DataTextureSceneModel````
+ * to build the simple table model, this time using geometry batching.
+ *
+ *  [![](http://xeokit.io/img/docs/sceneGraph.png)](/examples/#sceneRepresentation_DataTextureSceneModel_batching)
+ *
+ * * [[Run this example](/examples/#sceneRepresentation_DataTextureSceneModel_batching)]
+ *
+ * ````javascript
+ * import {Viewer, DataTextureSceneModel} from "xeokit-sdk.es.js";
+ *
+ * const viewer = new Viewer({
+ *     canvasId: "myCanvas",
+ *     transparent: true
+ * });
+ *
+ * viewer.scene.camera.eye = [-21.80, 4.01, 6.56];
+ * viewer.scene.camera.look = [0, -5.75, 0];
+ * viewer.scene.camera.up = [0.37, 0.91, -0.11];
+ *
+ * // Create a DataTextureSceneModel representing a table with four legs, using geometry batching
+ * const dataTextureSceneModel = new DataTextureSceneModel(viewer.scene, {
+ *     id: "table",
+ *     isModel: true,  // <--- Registers DataTextureSceneModel in viewer.scene.models
+ *     position: [0, 0, 0],
+ *     scale: [1, 1, 1],
+ *     rotation: [0, 0, 0]
+ * });
+ *
+ * // Red table leg
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "redLegMesh",
+ *
+ *     // Geometry arrays are same as for the earlier batching example
+ *     primitive: "triangles",
+ *     positions: [ 1, 1, 1, -1, 1, 1, -1, -1, 1, 1, -1, 1 ... ],
+ *     normals: [ 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, ... ],
+ *     indices: [ 0, 1, 2, 0, 2, 3, 4, 5, 6, 4, 6, 7, ... ],
+ *     position: [-4, -6, -4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [1, 0.3, 0.3],
+ *     origin: cfg.origin
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     id: "redLeg",
+ *     meshIds: ["redLegMesh"],
+ *     isObject: true // <---- Registers Entity by ID on viewer.scene.objects
+ * });
+ *
+ * // Green table leg
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "greenLegMesh",
+ *     primitive: "triangles",
+ *     primitive: "triangles",
+ *     positions: [ 1, 1, 1, -1, 1, 1, -1, -1, 1, 1, -1, 1 ... ],
+ *     normals: [ 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, ... ],
+ *     indices: [ 0, 1, 2, 0, 2, 3, 4, 5, 6, 4, 6, 7, ... ],
+ *     position: [4, -6, -4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [0.3, 1.0, 0.3],
+ *     origin: cfg.origin
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     id: "greenLeg",
+ *     meshIds: ["greenLegMesh"],
+ *     isObject: true // <---- Registers Entity by ID on viewer.scene.objects
+ * });
+ *
+ * // Blue table leg
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "blueLegMesh",
+ *     primitive: "triangles",
+ *     positions: [ 1, 1, 1, -1, 1, 1, -1, -1, 1, 1, -1, 1 ... ],
+ *     normals: [ 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, ... ],
+ *     indices: [ 0, 1, 2, 0, 2, 3, 4, 5, 6, 4, 6, 7, ... ],
+ *     position: [4, -6, 4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [0.3, 0.3, 1.0],
+ *     origin: cfg.origin
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     id: "blueLeg",
+ *     meshIds: ["blueLegMesh"],
+ *     isObject: true // <---- Registers Entity by ID on viewer.scene.objects
+ * });
+ *
+ * // Yellow table leg object
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "yellowLegMesh",
+ *     positions: [ 1, 1, 1, -1, 1, 1, -1, -1, 1, 1, -1, 1 ... ],
+ *     normals: [ 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, ... ],
+ *     indices: [ 0, 1, 2, 0, 2, 3, 4, 5, 6, 4, 6, 7, ... ],
+ *     position: [-4, -6, 4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [1.0, 1.0, 0.0],
+ *     origin: cfg.origin
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     id: "yellowLeg",
+ *     meshIds: ["yellowLegMesh"],
+ *     isObject: true // <---- Registers Entity by ID on viewer.scene.objects
+ * });
+ *
+ * // Purple table top
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "purpleTableTopMesh",
+ *     positions: [ 1, 1, 1, -1, 1, 1, -1, -1, 1, 1, -1, 1 ... ],
+ *     normals: [ 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, ... ],
+ *     indices: [ 0, 1, 2, 0, 2, 3, 4, 5, 6, 4, 6, 7, ... ],
+ *     position: [0, -3, 0],
+ *     scale: [6, 0.5, 6],
+ *     rotation: [0, 0, 0],
+ *     color: [1.0, 0.3, 1.0],
+ *     origin: cfg.origin
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     id: "purpleTableTop",
+ *     meshIds: ["purpleTableTopMesh"],
+ *     isObject: true // <---- Registers Entity by ID on viewer.scene.objects
+ * });
+ *
+ * // Finalize the DataTextureSceneModel.
+ *
+ * dataTextureSceneModel.finalize();
+ *
+ * // Find BigModelNodes by their model and object IDs
+ *
+ * // Get the whole table model
+ * const table = viewer.scene.models["table"];
+ *
+ * // Get some leg objects
+ * const redLeg = viewer.scene.objects["redLeg"];
+ * const greenLeg = viewer.scene.objects["greenLeg"];
+ * const blueLeg = viewer.scene.objects["blueLeg"];
+ * ````
+ *
+ * ## Classifying with Metadata
+ *
+ * In the previous examples, we used ````DataTextureSceneModel```` to build
+ * two versions of the same table model, to demonstrate geometry batching and geometry instancing.
+ *
+ * We'll now classify our {@link Entity}s with metadata. This metadata
+ * will work the same for both our examples, since they create the exact same structure of {@link Entity}s
+ * to represent their models and objects. The abstract Entity type is, after all, intended to provide an abstract interface through which differently-implemented scene content can be accessed uniformly.
+ *
+ * To create the metadata, we'll create a {@link MetaModel} for our model,
+ * with a {@link MetaObject} for each of it's objects. The MetaModel and MetaObjects
+ * get the same IDs as the {@link Entity}s that represent their model and objects within our scene.
+ *
+ * ```` javascript
+ * const furnitureMetaModel = viewer.metaScene.createMetaModel("furniture", {         // Creates a MetaModel in the MetaScene
+ *
+ *      "projectId": "myTableProject",
+ *      "revisionId": "V1.0",
+ *
+ *      "metaObjects": [
+ *          {                               // Creates a MetaObject in the MetaModel
+ *              "id": "table",
+ *              "name": "Table",            // Same ID as an object Entity
+ *              "type": "furniture",        // Arbitrary type, could be IFC type
+ *              "properties": {             // Arbitrary properties, could be IfcPropertySet
+ *                  "cost": "200"
+ *              }
+ *          },
+ *          {
+ *              "id": "redLeg",
+ *              "name": "Red table Leg",
+ *              "type": "leg",
+ *              "parent": "table",           // References first MetaObject as parent
+ *              "properties": {
+ *                  "material": "wood"
+ *              }
+ *          },
+ *          {
+ *              "id": "greenLeg",           // Node with corresponding id does not need to exist
+ *              "name": "Green table leg",  // and MetaObject does not need to exist for Node with an id
+ *              "type": "leg",
+ *              "parent": "table",
+ *              "properties": {
+ *                  "material": "wood"
+ *              }
+ *          },
+ *          {
+ *              "id": "blueLeg",
+ *              "name": "Blue table leg",
+ *              "type": "leg",
+ *              "parent": "table",
+ *              "properties": {
+ *                  "material": "wood"
+ *              }
+ *          },
+ *          {
+ *              "id": "yellowLeg",
+ *              "name": "Yellow table leg",
+ *              "type": "leg",
+ *              "parent": "table",
+ *              "properties": {
+ *                  "material": "wood"
+ *              }
+ *          },
+ *          {
+ *              "id": "tableTop",
+ *              "name": "Purple table top",
+ *              "type": "surface",
+ *              "parent": "table",
+ *              "properties": {
+ *                  "material": "formica",
+ *                  "width": "60",
+ *                  "depth": "60",
+ *                  "thickness": "5"
+ *              }
+ *          }
+ *      ]
+ *  });
+ * ````
+ *
+ * ## Querying Metadata
+ *
+ * Having created and classified our model (either the instancing or batching example), we can now find the {@link MetaModel}
+ * and {@link MetaObject}s using the IDs of their
+ * corresponding {@link Entity}s.
+ *
+ * ````JavaScript
+ * const furnitureMetaModel = scene.metaScene.metaModels["furniture"];
+ *
+ * const redLegMetaObject = scene.metaScene.metaObjects["redLeg"];
+ * ````
+ *
+ * In the snippet below, we'll log metadata on each {@link Entity} we click on:
+ *
+ * ````JavaScript
+ * viewer.scene.input.on("mouseclicked", function (coords) {
+ *
+ *      const hit = viewer.scene.pick({
+ *          canvasPos: coords
+ *      });
+ *
+ *      if (hit) {
+ *          const entity = hit.entity;
+ *          const metaObject = viewer.metaScene.metaObjects[entity.id];
+ *          if (metaObject) {
+ *              console.log(JSON.stringify(metaObject.getJSON(), null, "\t"));
+ *          }
+ *      }
+ *  });
+ * ````
+ *
+ * ## Metadata Structure
+ *
+ * The {@link MetaModel}
+ * organizes its {@link MetaObject}s in
+ * a tree that describes their structural composition:
+ *
+ * ````JavaScript
+ * // Get metadata on the root object
+ * const tableMetaObject = furnitureMetaModel.rootMetaObject;
+ *
+ * // Get metadata on the leg objects
+ * const redLegMetaObject = tableMetaObject.children[0];
+ * const greenLegMetaObject = tableMetaObject.children[1];
+ * const blueLegMetaObject = tableMetaObject.children[2];
+ * const yellowLegMetaObject = tableMetaObject.children[3];
+ * ````
+ *
+ * Given an {@link Entity}, we can find the object or model of which it is a part, or the objects that comprise it. We can also generate UI
+ * components from the metadata, such as the tree view demonstrated in [this demo](/examples/#BIMOffline_glTF_OTCConferenceCenter).
+ *
+ * This hierarchy allows us to express the hierarchical structure of a model while representing it in
+ * various ways in the 3D scene (such as with ````DataTextureSceneModel````, which
+ * has a non-hierarchical scene representation).
+ *
+ * Note also that a {@link MetaObject} does not need to have a corresponding
+ * {@link Entity} and vice-versa.
+ *
+ * # RTC Coordinates for Double Precision
+ *
+ * ````DataTextureSceneModel```` can emulate 64-bit precision on GPUs using relative-to-center (RTC) coordinates.
+ *
+ * Consider a model that contains many small objects, but with such large spatial extents that 32 bits of GPU precision (accurate to ~7 digits) will not be sufficient to render all of the the objects without jittering.
+ *
+ * To prevent jittering, we could spatially subdivide the objects into "tiles". Each tile would have a center position, and the positions of the objects within the tile would be relative to that center ("RTC coordinates").
+ *
+ * While the center positions of the tiles would be 64-bit values, the object positions only need to be 32-bit.
+ *
+ * Internally, when rendering an object with RTC coordinates, xeokit first temporarily translates the camera viewing matrix by the object's tile's RTC center, on the CPU, using 64-bit math.
+ *
+ * Then xeokit loads the viewing matrix into its WebGL shaders, where math happens at 32-bit precision. Within the shaders, the matrix is effectively down-cast to 32-bit precision, and the object's 32-bit vertex positions are transformed by the matrix.
+ *
+ * We see no jittering, because with RTC a detectable loss of GPU accuracy only starts happening to objects as they become very distant from the camera viewpoint, at which point they are too small to be discernible anyway.
+ *
+ * ## RTC Coordinates with Geometry Instancing
+ *
+ * To use RTC with ````DataTextureSceneModel```` geometry instancing, we specify an RTC center for the geometry via its ````origin```` parameter. Then ````DataTextureSceneModel```` assumes that all meshes that instance that geometry are within the same RTC coordinate system, ie. the meshes ````position```` and ````rotation```` properties are assumed to be relative to the geometry's ````origin````.
+ *
+ * For simplicity, our example's meshes all instance the same geometry. Therefore, our example model has only one RTC center.
+ *
+ * Note that the axis-aligned World-space boundary (AABB) of our model is ````[ -6, -9, -6, 1000000006, -2.5, 1000000006]````.
+ *
+ * [![](http://xeokit.io/img/docs/sceneGraph.png)](/examples/#sceneRepresentation_DataTextureSceneModel_batching)
+ *
+ * * [[Run this example](/examples/#sceneRepresentation_DataTextureSceneModel_instancing_origin)]
+ *
+ * ````javascript
+ * const origin = [100000000, 0, 100000000];
+ *
+ * dataTextureSceneModel.createGeometry({
+ *     id: "box",
+ *     primitive: "triangles",
+ *     positions: [ 1, 1, 1, -1, 1, 1, -1, -1, 1, 1, -1, 1 ... ],
+ *     normals: [ 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, ... ],
+ *     indices: [ 0, 1, 2, 0, 2, 3, 4, 5, 6, 4, 6, 7, ... ],
+ * });
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "leg1",
+ *     geometryId: "box",
+ *     position: [-4, -6, -4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [1, 0.3, 0.3],
+ *     origin: origin
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     meshIds: ["leg1"],
+ *     isObject: true
+ * });
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "leg2",
+ *     geometryId: "box",
+ *     position: [4, -6, -4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [0.3, 1.0, 0.3],
+ *     origin: origin
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     meshIds: ["leg2"],
+ *     isObject: true
+ * });
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "leg3",
+ *     geometryId: "box",
+ *     position: [4, -6, 4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [0.3, 0.3, 1.0],
+ *     origin: origin
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     meshIds: ["leg3"],
+ *     isObject: true
+ * });
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "leg4",
+ *     geometryId: "box",
+ *     position: [-4, -6, 4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [1.0, 1.0, 0.0],
+ *     origin: origin
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     meshIds: ["leg4"],
+ *     isObject: true
+ * });
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "top",
+ *     geometryId: "box",
+ *     position: [0, -3, 0],
+ *     scale: [6, 0.5, 6],
+ *     rotation: [0, 0, 0],
+ *     color: [1.0, 0.3, 1.0],
+ *     origin: origin
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     meshIds: ["top"],
+ *     isObject: true
+ * });
+ * ````
+ *
+ * ## RTC Coordinates with Geometry Batching
+ *
+ * To use RTC with ````DataTextureSceneModel```` geometry batching, we specify an RTC center (````origin````) for each mesh. For performance, we try to have as many meshes share the same value for ````origin```` as possible. Each mesh's ````positions````, ````position```` and ````rotation```` properties are assumed to be relative to ````origin````.
+ *
+ * For simplicity, the meshes in our example all share the same RTC center.
+ *
+ * The axis-aligned World-space boundary (AABB) of our model is ````[ -6, -9, -6, 1000000006, -2.5, 1000000006]````.
+ *
+ * [![](http://xeokit.io/img/docs/sceneGraph.png)](/examples/#sceneRepresentation_DataTextureSceneModel_batching)
+ *
+ * * [[Run this example](/examples/#sceneRepresentation_DataTextureSceneModel_batching_origin)]
+ *
+ * ````javascript
+ * const origin = [100000000, 0, 100000000];
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "leg1",
+ *     origin: origin, // This mesh's positions and transforms are relative to the RTC center
+ *     primitive: "triangles",
+ *     positions: [ 1, 1, 1, -1, 1, 1, -1, -1, 1, 1, -1, 1 ... ],
+ *     normals: [ 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, ... ],
+ *     indices: [ 0, 1, 2, 0, 2, 3, 4, 5, 6, 4, 6, 7, ... ],
+ *     position: [-4, -6, -4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [1, 0.3, 0.3]
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     meshIds: ["leg1"],
+ *     isObject: true
+ * });
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "leg2",
+ *     origin: origin, // This mesh's positions and transforms are relative to the RTC center
+ *     primitive: "triangles",
+ *     positions: [ 1, 1, 1, -1, 1, 1, -1, -1, 1, 1, -1, 1 ... ],
+ *     normals: [ 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, ... ],
+ *     indices: [ 0, 1, 2, 0, 2, 3, 4, 5, 6, 4, 6, 7, ... ],
+ *     position: [4, -6, -4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [0.3, 1.0, 0.3]
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     meshIds: ["leg2"],
+ *     isObject: true
+ * });
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "leg3",
+ *     origin: origin, // This mesh's positions and transforms are relative to the RTC center
+ *     primitive: "triangles",
+ *     positions: [ 1, 1, 1, -1, 1, 1, -1, -1, 1, 1, -1, 1 ... ],
+ *     normals: [ 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, ... ],
+ *     indices: [ 0, 1, 2, 0, 2, 3, 4, 5, 6, 4, 6, 7, ... ],
+ *     position: [4, -6, 4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [0.3, 0.3, 1.0]
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     meshIds: ["leg3"],
+ *     isObject: true
+ * });
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "leg4",
+ *     origin: origin, // This mesh's positions and transforms are relative to the RTC center
+ *     primitive: "triangles",
+ *     positions: [ 1, 1, 1, -1, 1, 1, -1, -1, 1, 1, -1, 1 ... ],
+ *     normals: [ 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, ... ],
+ *     indices: [ 0, 1, 2, 0, 2, 3, 4, 5, 6, 4, 6, 7, ... ],
+ *     position: [-4, -6, 4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [1.0, 1.0, 0.0]
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     meshIds: ["leg4"],
+ *     isObject: true
+ * });
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "top",
+ *     origin: origin, // This mesh's positions and transforms are relative to the RTC center
+ *     primitive: "triangles",
+ *     positions: [ 1, 1, 1, -1, 1, 1, -1, -1, 1, 1, -1, 1 ... ],
+ *     normals: [ 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, ... ],
+ *     indices: [ 0, 1, 2, 0, 2, 3, 4, 5, 6, 4, 6, 7, ... ],
+ *     position: [0, -3, 0],
+ *     scale: [6, 0.5, 6],
+ *     rotation: [0, 0, 0],
+ *     color: [1.0, 0.3, 1.0]
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     meshIds: ["top"],
+ *     isObject: true
+ * });
+ * ````
+ *
+ * ## Positioning at World-space coordinates
+ *
+ * To position a DataTextureSceneModel at given double-precision World coordinates, we can
+ * configure the ````origin```` of the DataTextureSceneModel itself. The ````origin```` is a double-precision
+ * 3D World-space position at which the DataTextureSceneModel will be located.
+ *
+ * Note that ````position```` is a single-precision offset relative to ````origin````.
+ *
+ * ````javascript
+ * const origin = [100000000, 0, 100000000];
+ *
+ * const dataTextureSceneModel = new DataTextureSceneModel(viewer.scene, {
+ *     id: "table",
+ *     isModel: true,
+ *     origin: origin, // Everything in this DataTextureSceneModel is relative to this RTC center
+ *     position: [0, 0, 0],
+ *     scale: [1, 1, 1],
+ *     rotation: [0, 0, 0]
+ * });
+ *
+ * dataTextureSceneModel.createGeometry({
+ *     id: "box",
+ *     primitive: "triangles",
+ *     positions: [ 1, 1, 1, -1, 1, 1, -1, -1, 1, 1, -1, 1 ... ],
+ *     normals: [ 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, ... ],
+ *     indices: [ 0, 1, 2, 0, 2, 3, 4, 5, 6, 4, 6, 7, ... ],
+ * });
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "leg1",
+ *     geometryId: "box",
+ *     position: [-4, -6, -4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [1, 0.3, 0.3]
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     meshIds: ["leg1"],
+ *     isObject: true
+ * });
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "leg2",
+ *     geometryId: "box",
+ *     position: [4, -6, -4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [0.3, 1.0, 0.3]
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     meshIds: ["leg2"],
+ *     isObject: true
+ * });
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "leg3",
+ *     geometryId: "box",
+ *     position: [4, -6, 4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [0.3, 0.3, 1.0]
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     meshIds: ["leg3"],
+ *     isObject: true
+ * });
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "leg4",
+ *     geometryId: "box",
+ *     position: [-4, -6, 4],
+ *     scale: [1, 3, 1],
+ *     rotation: [0, 0, 0],
+ *     color: [1.0, 1.0, 0.0]
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     meshIds: ["leg4"],
+ *     isObject: true
+ * });
+ *
+ * dataTextureSceneModel.createMesh({
+ *     id: "top",
+ *     geometryId: "box",
+ *     position: [0, -3, 0],
+ *     scale: [6, 0.5, 6],
+ *     rotation: [0, 0, 0],
+ *     color: [1.0, 0.3, 1.0]
+ * });
+ *
+ * dataTextureSceneModel.createEntity({
+ *     meshIds: ["top"],
+ *     isObject: true
+ * });
+ * ````
  *
  * @implements {Drawable}
  * @implements {Entity}
+ * @implements {SceneModel}
+ * @implements {ViewFrustumCullingModel}
  */
-class DataTextureSceneModel extends Component {
+export class DataTextureSceneModel extends Component {
 
     /**
      * @constructor
@@ -54,13 +958,13 @@ class DataTextureSceneModel extends Component {
      * @param {Boolean} [cfg.edges=false] Indicates if the DataTextureSceneModel's edges are initially emphasized.
      * @param {Number[]} [cfg.colorize=[1.0,1.0,1.0]] DataTextureSceneModel's initial RGB colorize color, multiplies by the rendered fragment colors.
      * @param {Number} [cfg.opacity=1.0] DataTextureSceneModel's initial opacity factor, multiplies by the rendered fragment alpha.
-     * @param {Number} [cfg.backfaces=false] When we set this ````true````, then we force rendering of backfaces for this DataTextureModel. When
+     * @param {Number} [cfg.backfaces=false] When we set this ````true````, then we force rendering of backfaces for this DataTextureSceneModel. When
      * we leave this ````false````, then we allow the Viewer to decide when to render backfaces. In that case, the
      * Viewer will hide backfaces on watertight meshes, show backfaces on open meshes, and always show backfaces on meshes when we slice them open with {@link SectionPlane}s.
-     * @param {Boolean} [cfg.saoEnabled=true] Indicates if Scalable Ambient Obscurance (SAO) will apply to this DataTextureModel. SAO is configured by the Scene's {@link SAO} component.
+     * @param {Boolean} [cfg.saoEnabled=true] Indicates if Scalable Ambient Obscurance (SAO) will apply to this DataTextureSceneModel. SAO is configured by the Scene's {@link SAO} component.
      * @param {Number} [cfg.edgeThreshold=10] When xraying, highlighting, selecting or edging, this is the threshold angle between normals of adjacent triangles, below which their shared wireframe edge is not drawn.
-     * @param {Number} [cfg.targetLodFps] Optional target LoD FPS. When provided, will enable LoD culling for this DataTextureModel, with the given target FPS.
-     * @param {Boolean} [cfg.enableViewFrustumCulling=false] When true, will enable view frustum culling for the objects within this DataTextureModel.
+     * @param {Number} [cfg.targetLodFps] Optional target LoD FPS. When provided, will enable LoD culling for this DataTextureSceneModel, with the given target FPS.
+     * @param {Boolean} [cfg.enableViewFrustumCulling=false] When true, will enable view frustum culling for the objects within this DataTextureSceneModel.
      * @param {Boolean} [cfg.disableVertexWelding] Disable vertex welding when loading geometry into the GPU. Default is ```false```.
      * @param {Boolean} [cfg.disableIndexRebucketing] Disable index rebucketing when loading geometry into the GPU. Default is ```false```.
      */
@@ -101,6 +1005,8 @@ class DataTextureSceneModel extends Component {
 
         this._aabb = math.collapseAABB3();
         this._aabbDirty = false;
+        this._currentLayers = {};
+        this._layers = {};
 
         /**
          * @type {Array<TrianglesDataTextureLayer>}
@@ -111,11 +1017,6 @@ class DataTextureSceneModel extends Component {
          * @type {Array<DataTextureSceneModelNode>}
          */
         this._nodeList = [];
-
-        /**
-         * @type {TrianglesDataTextureLayer}
-         */
-        this._currentDataTextureLayer = null;
 
         this._instancingGeometries = {};
 
@@ -199,7 +1100,6 @@ class DataTextureSceneModel extends Component {
 
         /** @private */
         this._numTriangles = 0;
-
         this._edgeThreshold = cfg.edgeThreshold || 10;
 
         this.visible = cfg.visible;
@@ -953,8 +1853,6 @@ class DataTextureSceneModel extends Component {
      * @param {Number[]} [cfg.indices] Array of indices. Not required for `points` primitives.
      * @param {Number[]} [cfg.edgeIndices] Array of edge line indices. Used only for Required for 'triangles' primitives. These are automatically generated internally if not supplied, using the ````edgeThreshold```` given to the ````DataTextureSceneModel```` constructor.
      * @param {Number[]} [cfg.positionsDecodeMatrix] A 4x4 matrix for decompressing ````positionsCompresse````.
-     * @param {Number[]} [cfg.origin] Optional geometry origin, relative to {@link DataTextureSceneModel#origin}. When this is given, then every mesh created with {@link DataTextureSceneModel#createMesh} that uses this geometry will
-     * be transformed relative to this origin.
      */
     createGeometry(cfg) {
         const geometryId = cfg.id;
@@ -985,8 +1883,6 @@ class DataTextureSceneModel extends Component {
             math.expandAABB3Points3(aabb, cfg.positions);
             cfg.positionsCompressed = quantizePositions(cfg.positions, aabb, cfg.positionsDecodeMatrix)
         }
-        const cfgOrigin = cfg.origin || cfg.rtcCenter;
-        const origin = (cfgOrigin) ? math.addVec3(this._origin, cfgOrigin, tempVec3a) : this._origin;
         switch (primitive) {
             case "triangles":
             case "solid":
@@ -995,7 +1891,7 @@ class DataTextureSceneModel extends Component {
                     this.error(`Config missing: indices - required for primitive type "${primitive}"`);
                     return;
                 }
-                const geometryCfg = utils.apply({origin, layerIndex: 0,}, cfg);
+                const geometryCfg = utils.apply({layerIndex: 0}, cfg);
                 if (!geometryCfg.edgeIndices) {
                     geometryCfg.edgeIndices = buildEdgeIndices(cfg.positionsCompressed, cfg.indices, null, 5.0);
                 }
@@ -1062,7 +1958,7 @@ class DataTextureSceneModel extends Component {
                 return;
             }
             if (cfg.primitive !== "points" && !cfg.indices) {
-                this.error(`Config missing: indices - required for ${cfg.primitive} primitives`);
+                this.error(`Config missing: indices - required with ${cfg.primitive} primitives`);
                 return;
             }
             if (cfg.positionsCompressed && !cfg.positionsDecodeMatrix) {
@@ -1081,45 +1977,28 @@ class DataTextureSceneModel extends Component {
             preparedGeometryCfg = this._preparedInstancingGeometries[geometryId];
             cfg.geometryId = geometryId;
         }
-        const origin = cfg.origin;
+        const origin = (cfg.origin) ? math.addVec3(this._origin, cfg.origin, tempVec3a) : this._origin;
         cfg.color = cfg.color ? cfg.color.slice() : [255, 255, 255];
         if (this._vfcManager && !this._vfcManager.finalized) {
             this._vfcManager.addMesh(cfg); // To be created in #finalize()
             return;
         }
-        let layer = this._currentDataTextureLayer;
-        if (null !== layer && !layer.canCreatePortion(preparedGeometryCfg, null)) {
-            layer.finalize();
-            delete this._currentDataTextureLayer;
-            layer = null;
-        }
-        if (!layer) {
-            layer = new TrianglesDataTextureLayer(this, {layerIndex: 0, origin}); // layerIndex is set in #finalize()
-            this._layerList.push(layer);
-            this._currentDataTextureLayer = layer;
-        }
-
+        let layer = this._getLayer(origin, preparedGeometryCfg);
         const color = (cfg.color) ? new Uint8Array([Math.floor(cfg.color[0] * 255), Math.floor(cfg.color[1] * 255), Math.floor(cfg.color[2] * 255)]) : [255, 255, 255];
         const opacity = (cfg.opacity !== undefined && cfg.opacity !== null) ? Math.floor(cfg.opacity * 255) : 255;
         const metallic = (cfg.metallic !== undefined && cfg.metallic !== null) ? Math.floor(cfg.metallic * 255) : 0;
         const roughness = (cfg.roughness !== undefined && cfg.roughness !== null) ? Math.floor(cfg.roughness * 255) : 255;
-
         const mesh = new DataTextureSceneModelMesh(this, id, color, opacity);
-
         const pickId = mesh.pickId;
         const a = pickId >> 24 & 0xFF;
         const b = pickId >> 16 & 0xFF;
         const g = pickId >> 8 & 0xFF;
         const r = pickId & 0xFF;
-
         const pickColor = new Uint8Array([r, g, b, a]); // Quantized pick color
         const aabb = math.collapseAABB3();
-
         preparedGeometryCfg.solid = preparedGeometryCfg.primitive === "solid";
-
         let meshMatrix;
-        let worldMatrix = this._worldMatrixNonIdentity ? this._worldMatrix : null;
-
+        const worldMatrix = this._worldMatrixNonIdentity ? this._worldMatrix : null;
         if (cfg.matrix) {
             meshMatrix = cfg.matrix;
         } else {
@@ -1129,7 +2008,6 @@ class DataTextureSceneModel extends Component {
             math.eulerToQuaternion(rotation, "XYZ", defaultQuaternion);
             meshMatrix = math.composeMat4(position, defaultQuaternion, scale, math.mat4());
         }
-
         const portionId = layer.createPortion(preparedGeometryCfg, {
             origin,
             geometryId,
@@ -1142,14 +2020,9 @@ class DataTextureSceneModel extends Component {
             aabb,
             pickColor
         });
-
         math.expandAABB3(this._aabb, aabb);
-
         this._numTriangles += preparedGeometryCfg.indices.length / 3;
         mesh.numTriangles = preparedGeometryCfg.indices.length / 3;
-
-        mesh.origin = preparedGeometryCfg.origin;
-        math.expandAABB3(this._aabb, aabb);
         this.numGeometries++;
         mesh.origin = origin;
         // noinspection JSConstantReassignment
@@ -1158,8 +2031,26 @@ class DataTextureSceneModel extends Component {
         mesh._portionId = portionId;
         // noinspection JSConstantReassignment
         mesh.aabb = aabb;
-
         this._meshes[id] = mesh;
+    }
+
+    _getLayer(origin, preparedGeometryCfg) {
+        const layerId = `${origin[0]}.${origin[1]}.${origin[2]}`;
+        let layer = this._currentLayers[layerId];
+        if (layer) {
+            if (!layer.canCreatePortion(preparedGeometryCfg, null)) {
+                layer.finalize();
+                delete this._currentLayers[layerId];
+                layer = null;
+            } else {
+                return layer;
+            }
+        }
+        layer = new TrianglesDataTextureLayer(this, {layerIndex: 0, origin}); // layerIndex is set in #finalize()
+        this._currentLayers[layerId] = layer;
+        this._layers[layerId] = layer;
+        this._layerList.push(layer);
+        return layer;
     }
 
     /**
@@ -1170,7 +2061,6 @@ class DataTextureSceneModel extends Component {
      * @param {Object} cfg Entity configuration.
      * @param {String} cfg.id Optional ID for the new Entity. Must not clash with any existing components within the {@link Scene}.
      * @param {String[]} cfg.meshIds IDs of one or more meshes created previously with {@link DataTextureSceneModel@createMesh}.
-
      * @param {Boolean} [cfg.isObject] Set ````true```` if the {@link Entity} represents an object, in which case it will be registered by {@link Entity#id} in {@link Scene#objects} and can also have a corresponding {@link MetaObject} with matching {@link MetaObject#id}, registered by that ID in {@link MetaScene#metaObjects}.
      * @param {Boolean} [cfg.visible=true] Indicates if the Entity is initially visible.
      * @param {Boolean} [cfg.culled=false] Indicates if the Entity is initially culled from view.
@@ -1190,7 +2080,7 @@ class DataTextureSceneModel extends Component {
         if (id === undefined) {
             id = math.createUUID();
         } else if (this.scene.components[id]) {
-            this.error("Scene already has a Component with this ID: " + id + " - will assign random ID");
+            this.error("Scene already has a Component with this ID: " + id + " - will assign random ID to new Entity");
             id = math.createUUID();
         }
         const meshIds = cfg.meshIds;
@@ -1273,16 +2163,16 @@ class DataTextureSceneModel extends Component {
         }
         if (this._vfcManager) {
             this._vfcManager.finalize(() => {
-                if (!this._currentDataTextureLayer) {
-                    return;
+                for (let layerId in this._currentLayers) {
+                    this._currentLayers[layerId].finalize();
+                    delete this._currentLayers[layerId];
                 }
-                this._currentDataTextureLayer.finalize();
-                delete this._currentDataTextureLayer;
-                this._currentDataTextureLayer = null;
             });
-        }
-        if (this._currentDataTextureLayer) {
-            this._currentDataTextureLayer.finalize();
+        } else {
+            for (let layerId in this._currentLayers) {
+                this._currentLayers[layerId].finalize();
+                delete this._currentLayers[layerId];
+            }
         }
         for (let i = 0, len = this._nodeList.length; i < len; i++) {
             const node = this._nodeList[i];
@@ -1359,7 +2249,9 @@ class DataTextureSceneModel extends Component {
         }
     }
 
-    /** @private */
+    /**
+     * @private
+     */
     _getActiveSectionPlanesForLayer(layer) {
         const renderFlags = this.renderFlags;
         const sectionPlanes = this.scene._sectionPlanesState.sectionPlanes;
@@ -1381,23 +2273,17 @@ class DataTextureSceneModel extends Component {
 
     /** @private */
     _updateRenderFlags() {
-
         if (this.numVisibleLayerPortions === 0) {
             return;
         }
-
         if (this.numCulledLayerPortions === this.numPortions) {
             return;
         }
-
         const renderFlags = this.renderFlags;
-
         renderFlags.colorOpaque = (this.numTransparentLayerPortions < this.numPortions);
-
         if (this.numTransparentLayerPortions > 0) {
             renderFlags.colorTransparent = true;
         }
-
         if (this.numXRayedLayerPortions > 0) {
             const xrayMaterial = this.scene.xrayMaterial._state;
             if (xrayMaterial.fill) {
@@ -1415,7 +2301,6 @@ class DataTextureSceneModel extends Component {
                 }
             }
         }
-
         if (this.numEdgesLayerPortions > 0) {
             const edgeMaterial = this.scene.edgeMaterial._state;
             if (edgeMaterial.edges) {
@@ -1425,7 +2310,6 @@ class DataTextureSceneModel extends Component {
                 }
             }
         }
-
         if (this.numSelectedLayerPortions > 0) {
             const selectedMaterial = this.scene.selectedMaterial._state;
             if (selectedMaterial.fill) {
@@ -1443,7 +2327,6 @@ class DataTextureSceneModel extends Component {
                 }
             }
         }
-
         if (this.numHighlightedLayerPortions > 0) {
             const highlightMaterial = this.scene.highlightMaterial._state;
             if (highlightMaterial.fill) {
@@ -1573,9 +2456,7 @@ class DataTextureSceneModel extends Component {
         }
     }
 
-    /**
-     * @private
-     */
+    /** @private */
     drawOcclusion(frameCtx) {
         if (this.numVisibleLayerPortions === 0) {
             return;
@@ -1587,9 +2468,7 @@ class DataTextureSceneModel extends Component {
         }
     }
 
-    /**
-     * @private
-     */
+    /** @private */
     drawShadow(frameCtx) {
         if (this.numVisibleLayerPortions === 0) {
             return;
@@ -1690,7 +2569,6 @@ class DataTextureSceneModel extends Component {
     }
 }
 
-
 /**
  * This function applies two steps to the provided mesh geometry data:
  *
@@ -1710,6 +2588,8 @@ class DataTextureSceneModel extends Component {
  *
  * @param {object} geometryCfg The mesh information containing `.positions`, `.indices`, `.edgeIndices` arrays.
  *
+ * @param enableVertexWelding
+ * @param enableIndexRebucketing
  * @returns {object} The mesh information enrichened with `.preparedBuckets` key.
  */
 function prepareMeshGeometry(geometryCfg, enableVertexWelding, enableIndexRebucketing) {
@@ -1753,5 +2633,3 @@ function prepareMeshGeometry(geometryCfg, enableVertexWelding, enableIndexRebuck
 
     return geometryCfg;
 }
-
-export {DataTextureSceneModel};
